[xrhome] Factor out useProjectConfigStatus
